updater module now more failsafe
if errors occur in updater module, the operator code won't crash - effectively jsut disables the updater and allows the rest of the addon to work still. Also added additional printout to specify active addon so if mutliple addons usse updater code, it is clear which printouts relate to which addons.