Readability update & showing GitLab ssl action
Now that GitLab connections over ssl have followed a similar path as GitHub in terms of minimum SSL library requirements, OSX users connecting to GitLab repos will see errors. This updates, similar to the previous time this occured, will ensure that a clearer message is given to the user (and to manually check for updates, with a quick link to do so). This update also improves various readability and some coding styles. Some additions are also helpful for integration of this code into a more generalized addon manager.